+++ Sources/AkuoCore/AkuoCoreVersion.swift
@@ -1,4 +1,4 @@
 public enum AkuoCoreVersion {
     public static let current = "0.4.0"
-    public static let build = "13"
+    public static let build = "12"
 }